requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.2.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
dotenv==0.9.9
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.3.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
pytest-django==4.11.1
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
